wip: db transaction for purchasing stock
