+++ backend/db/sqlc/store.go
@@ -1,7 +1,9 @@
 package db
 
 import (
+	"context"
 	"database/sql"
+	"fmt"
 )
 
 // Store provides all functions to execute db queries and transactions
@@ -18,7 +20,39 @@ func NewStore(db *sql.DB) *Store {
 	}
 }
 
-//buying stock
+// ExecTx executes a function within a database transaction
+func (store *Store) execTx(ctx context.Context, fn func(*Queries) error) error {
+	tx, err := store.db.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+
+	q := New(tx)
+	err = fn(q)
+	if err != nil {
+		if rbErr := tx.Rollback(); rbErr != nil {
+			return fmt.Errorf("tx err: %v, rb err: %v", err, rbErr)
+		}
+		return err
+	}
+
+	return tx.Commit()
+}
+
+type BuyStockParams struct {
+	AccountID int64  `json:"account_id"`
+	StockName string `json:"stock_name"`
+	Amount    int64  `json:"amount"`
+}
+
+type BuyStockTxResult struct {
+}
+
+// BuyStockTx performs a money subtraction from an account and adds stock to an account
+// update stock amount in user stock and update user dollar and cents in a single transaction
+func (store *Store) BuyStockTx(ctx context.Context, arg BuyStockParams) (BuyStockTxResult, error) {
+
+}
 
 //selling stock
